+++ src/components/modals/ColumnSections.ts
@@ -10,6 +10,11 @@ import { DataTypes } from "helpers/Constants";
 import { AbstractChain, AbstractHandler } from "patterns/AbstractFactoryChain";
 
 class BehaviorSetttingsSection extends AbstractChain<ColumnHandlerResponse> {
+    private dataType: string = DataTypes.TEXT;
+    protected runBefore(columnHandlerResponse: ColumnHandlerResponse): ColumnHandlerResponse {
+        this.dataType = columnHandlerResponse.column.dataType;
+        return columnHandlerResponse;
+    }
     protected customHandle(columnHandlerResponse: ColumnHandlerResponse): ColumnHandlerResponse {
         const behavior_section = columnHandlerResponse.containerEl.createDiv("column-section-container-behavior");
         add_setting_header(behavior_section, "Behavior", "h3");
@@ -17,9 +22,15 @@ class BehaviorSetttingsSection extends AbstractChain<ColumnHandlerResponse> {
         return columnHandlerResponse;
     }
     protected getHandlers(): AbstractHandler<ColumnHandlerResponse>[] {
-        return [
-            new InlineToggleHandler()
-        ]
+        const particularHandlers: ColumnHandler[] = [];
+        switch (this.dataType) {
+            case DataTypes.TASK:
+                // do nothing
+                break;
+            default:
+                particularHandlers.push(new InlineToggleHandler());
+        }
+        return particularHandlers;
     }
 }
 export const behavior_settings_section = new BehaviorSetttingsSection();
@@ -28,6 +39,12 @@ export const behavior_settings_section = new BehaviorSetttingsSection();
  * Every column type has a different behavior section
  */
 class ParticularSetttingsSection extends AbstractChain<ColumnHandlerResponse> {
+    private dataType: string = DataTypes.TEXT;
+    protected runBefore(columnHandlerResponse: ColumnHandlerResponse): ColumnHandlerResponse {
+        this.dataType = columnHandlerResponse.column.dataType;
+        return columnHandlerResponse;
+    }
+
     protected customHandle(columnHandlerResponse: ColumnHandlerResponse): ColumnHandlerResponse {
         const particular_section = columnHandlerResponse.containerEl.createDiv("column-section-container-particular");
         // title of the section
@@ -35,29 +52,23 @@ class ParticularSetttingsSection extends AbstractChain<ColumnHandlerResponse> {
         columnHandlerResponse.containerEl = particular_section;
         return columnHandlerResponse;
     }
-    protected getHandlers(option?: string): AbstractHandler<ColumnHandlerResponse>[] {
-        return [
-            new InlineToggleHandler()
-        ]
+    protected getHandlers(): AbstractHandler<ColumnHandlerResponse>[] {
+        const particularHandlers: ColumnHandler[] = [];
+        switch (this.dataType) {
+            case DataTypes.TEXT:
+                particularHandlers.push(new MediaToggleHandler());
+                particularHandlers.push(new MediaDimensionsHandler());
+                break;
+            case DataTypes.SELECT:
+            case DataTypes.TAGS:
+                particularHandlers.push(new SelectedColumnOptionsHandler());
+                break;
+            case DataTypes.TASK:
+                particularHandlers.push(new HideCompletedTaskToggleHandler());
+            default:
+                break;
+        }
+        return particularHandlers;
     }
 }
 export const particular_settings_section = new ParticularSetttingsSection();
-
-function addParticularInputSettings(dataType: string): ColumnHandler[] {
-    const particularHandlers: ColumnHandler[] = [];
-    switch (dataType) {
-        case DataTypes.TEXT:
-            particularHandlers.push(new MediaToggleHandler());
-            particularHandlers.push(new MediaDimensionsHandler());
-            break;
-        case DataTypes.SELECT:
-        case DataTypes.TAGS:
-            particularHandlers.push(new SelectedColumnOptionsHandler());
-            break;
-        case DataTypes.TASK:
-            particularHandlers.push(new HideCompletedTaskToggleHandler());
-        default:
-            break;
-    }
-    return particularHandlers;
-}